# openapi_schema.json
{"openapi":"3.1.0","info":{"title":"Config","description":"Orquestador de agentes de IA con OpenAI","version":"1.0.0"},"servers":[{"url":"REEMPLAZAR_CON_URL_NGROK","description":"Servidor local expuesto via ngrok"}],"paths":{"/gpt/task/execute":{"post":{"tags":["GPT Actions"],"summary":"Execute Agent Task","description":"ENDPOINT PARA ACTIONS: Ejecutar tarea con agente específico\n\nUse this in your GPT Actions to delegate tasks to specialists\n\nExample from Andrés (Director):\n{\n    \"agent_id\": \"adrian_datos\",\n    \"task\": \"Analiza las ventas del último mes por país\",\n    \"context\": {\"period\": \"last_month\"}\n}","operationId":"execute_agent_task_gpt_task_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gpt/director/coordinate":{"post":{"tags":["GPT Actions"],"summary":"Director Coordinate","description":"ENDPOINT PARA ACTIONS: Director (Andrés) coordina múltiples agentes\n\nUse this when you need complex coordination\n\nExample:\n{\n    \"request\": \"Necesito análisis de ventas Q4 y propuesta de estrategia\",\n    \"required_agents\": [\"adrian_datos\", \"bruno_estrategia\"],\n    \"context\": {}\n}","operationId":"director_coordinate_gpt_director_coordinate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectorRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gpt/agents/list":{"get":{"tags":["GPT Actions"],"summary":"List All Agents","description":"ENDPOINT PARA ACTIONS: Listar todos los agentes disponibles\n\nUse this in your GPT to understand who's available in the team","operationId":"list_all_agents_gpt_agents_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/gpt/team/analyze":{"post":{"tags":["GPT Actions"],"summary":"Team Analyze","description":"ENDPOINT PARA ACTIONS: Análisis coordinado del equipo\n\nMultiple agents analyze the same problem from different angles\n\nExample:\n{\n    \"project\": \"Expansion Q1 2024\",\n    \"objective\": \"Analizar viabilidad de expansión a nuevos mercados\",\n    \"agents\": [\"adrian_datos\", \"leo_partners\", \"bruno_estrategia\"],\n    \"context\": {}\n}","operationId":"team_analyze_gpt_team_analyze_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TeamCoordinationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gpt/workflow/execute":{"post":{"tags":["GPT Actions"],"summary":"Execute Workflow","description":"ENDPOINT PARA ACTIONS: Ejecutar workflow completo (pipeline)\n\nAgents work in sequence, each one using previous output\n\nExample:\n{\n    \"workflow_name\": \"Marketing Campaign Analysis\",\n    \"steps\": [\n        {\"agent\": \"adrian_datos\", \"task\": \"Analiza data histórica de campañas\"},\n        {\"agent\": \"diego_automatizacion\", \"task\": \"Propón automatizaciones basadas en análisis\"},\n        {\"agent\": \"andres_director\", \"task\": \"Revisa y da aprobación final\"}\n    ]\n}","operationId":"execute_workflow_gpt_workflow_execute_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Request"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gpt/health/team":{"get":{"tags":["GPT Actions"],"summary":"Team Health","description":"ENDPOINT PARA ACTIONS: Verificar salud del equipo\n\nCheck that all agents are available","operationId":"team_health_gpt_health_team_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"summary":"Root","description":"Endpoint raíz - verificar que la API está funcionando","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/agents":{"get":{"summary":"List Agents","description":"Listar todos los agentes disponibles","operationId":"list_agents_agents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/execute/simple":{"post":{"summary":"Execute Simple Task","description":"Ejecutar una tarea simple con un agente específico\n\nEjemplo:\n{\n    \"agent_id\": \"analyst\",\n    \"prompt\": \"Analiza este conjunto de datos...\",\n    \"context\": {}\n}","operationId":"execute_simple_task_execute_simple_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/execute/pipeline":{"post":{"summary":"Execute Pipeline","description":"Ejecutar un pipeline donde la salida de un agente es entrada del siguiente\n\nEjemplo:\n{\n    \"agents_sequence\": [\"analyst\", \"processor\", \"coordinator\"],\n    \"initial_prompt\": \"Analiza estos datos y crea un reporte...\"\n}","operationId":"execute_pipeline_execute_pipeline_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PipelineRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/execute/parallel":{"post":{"summary":"Execute Parallel Analysis","description":"Ejecutar el mismo prompt en múltiples agentes en paralelo\n\nEjemplo:\n{\n    \"prompt\": \"¿Cuál es tu análisis sobre esto?\",\n    \"agents\": [\"analyst\", \"processor\"]\n}","operationId":"execute_parallel_analysis_execute_parallel_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParallelAnalysisRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/history":{"get":{"summary":"Get Execution History","description":"Obtener el historial de todas las ejecuciones","operationId":"get_execution_history_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Verificar que la API está saludable","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AgentRequest":{"properties":{"agent_id":{"type":"string","title":"Agent Id"},"prompt":{"type":"string","title":"Prompt"},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context"}},"type":"object","required":["agent_id","prompt"],"title":"AgentRequest","description":"Solicitud a un agente"},"DirectorRequest":{"properties":{"request":{"type":"string","title":"Request"},"required_agents":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Required Agents"},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context"}},"type":"object","required":["request"],"title":"DirectorRequest","description":"Solicitud al Director (Andrés) para orquestar"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ParallelAnalysisRequest":{"properties":{"prompt":{"type":"string","title":"Prompt"},"agents":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Agents"}},"type":"object","required":["prompt"],"title":"ParallelAnalysisRequest","description":"Solicitud para análisis en paralelo"},"PipelineRequest":{"properties":{"agents_sequence":{"items":{"type":"string"},"type":"array","title":"Agents Sequence"},"initial_prompt":{"type":"string","title":"Initial Prompt"}},"type":"object","required":["agents_sequence","initial_prompt"],"title":"PipelineRequest","description":"Solicitud para ejecutar un pipeline de agentes"},"TaskRequest":{"properties":{"agent_id":{"type":"string","title":"Agent Id"},"task":{"type":"string","title":"Task"},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context"}},"type":"object","required":["agent_id","task"],"title":"TaskRequest","description":"Solicitud de tarea para un agente especializado"},"TeamCoordinationRequest":{"properties":{"project":{"type":"string","title":"Project"},"objective":{"type":"string","title":"Objective"},"agents":{"items":{"type":"string"},"type":"array","title":"Agents"},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context"}},"type":"object","required":["project","objective","agents"],"title":"TeamCoordinationRequest","description":"Solicitud de coordinación entre múltiples agentes"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}